hf_inference_backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.3.1
numpy==1.26.4
pandas==2.2.2
huggingface_hub==0.24.5
umap-learn==0.5.6
matplotlib==3.8.4
python-multipart==0.0.9
